Refactor scraper logic to single directory
