+++ app/commands.py
@@ -8,12 +8,15 @@ from flask.cli import with_appcontext
 
 from app.database import db
 from app.models import Person
-from app.other_scraper.institutions import import_institutions_handler
-from app.other_scraper.legal_areas import import_legal_areas_handler
-from app.other_scraper.procedure_types import import_procedure_types_handler
-from app.people_scraper.extract import enrich_people_handler, import_people_handler
-from app.verdict_scraper.extract import enrich_verdicts_handler, import_verdicts_handler
-from app.verdict_scraper.from_files import import_verdicts_from_files_handler
+from app.scraper.other.institutions import import_institutions_handler
+from app.scraper.other.legal_areas import import_legal_areas_handler
+from app.scraper.other.procedure_types import import_procedure_types_handler
+from app.scraper.people.extract import enrich_people_handler, import_people_handler
+from app.scraper.verdicts.extract import (
+    enrich_verdicts_handler,
+    import_verdicts_handler,
+)
+from app.scraper.verdicts.from_files import import_verdicts_from_files_handler
 
 
 @click.command("placeholder")
