+++ app/scraper/other/legal_areas.py
@@ -2,7 +2,7 @@ import requests
 from flask import current_app
 
 from app.models import LegalArea
-from app.verdict_scraper.soup_parsing import safe_find_text, to_soup
+from app.scraper.verdicts.soup_parsing import safe_find_text, to_soup
 
 
 def transform_legal_area_xml_to_dict(area):
