+++ app/scraper/other/procedure_types.py
@@ -2,7 +2,7 @@ import requests
 from flask import current_app
 
 from app.models import ProcedureType
-from app.verdict_scraper.soup_parsing import safe_find_text, to_soup
+from app.scraper.verdicts.soup_parsing import safe_find_text, to_soup
 
 
 def transform_procedure_type_xml_to_dict(soup):
