+++ app/scraper/people/config.py
