+++ app/scraper/verdicts/config.py
