+++ app/scraper/verdicts/soup_parsing.py
