+++ app/scraper/verdicts/utils.py
